fix warning in 32bit ; use of COUNTER3 macro

+++ src/spdr-thread-posix.c
@@ -4,5 +4,5 @@
 
 uint64_t uu_spdr_get_tid()
 {
-	return (uint64_t) pthread_self();
+	return (intptr_t) pthread_self();
 }
